requirements.txt
arxiv
PyMuPDF
sentence-transformers
faiss-cpu
transformers
torch
numpy
pandas
tqdm
rouge-score
nltk
scikit-learn
matplotlib
seaborn
requests
python-dotenv
google-genai
streamlit
rank-bm25
tf-keras
